refactor: improve code readability and organization
- Extract inline HTML templates to dedicated template package
- Reorganize ApplicationContext struct with logical groupings
- Improve import organization with proper grouping
- Replace large inline constants with external template files
- Enhance template styling with improved UX (loading spinner, better CSS)
- Maintain exact functionality while improving maintainability

Files reorganized:
- cmd/server/main.go: Better structure and organization
- internal/auth/discord.go: Cleaner template usage
- internal/templates/: New package for HTML templates
